pyproject.toml
[tool.poetry]
name = "telegrambot01"
version = "0.1.0"
description = ""
authors = ["Сергей Сушков <sushkov_sv@ursip.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
aiohttp = "^3.9.3"
aiosignal = "^1.3.1"
async-timeout = "^4.0.3"
attrs = "^23.2.0"
babel = "^2.14.0"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"
cachetools = "^5.3.3"
certifi = "^2024.2.2"
charset-normalizer = "^3.3.2"
deprecation = "^2.1.0"
frozenlist = "^1.4.1"
grpcio = "^1.62.1"
idna = "^3.6"
magic-filter = "^1.0.12"
multidict = "^6.0.5"
numpy = "^1.26.4"
packaging = "^24.0"
pandas = "^2.2.1"
pillow = "^10.2.0"
protobuf = "^4.25.1"
psycopg2 = "^2.9.9"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.28"
tinkoff = "^0.1.1"
greenlet = "^3.0.3"
tinkoff-investments = "^0.2.0b96"
tzdata = "^2024.1"
urllib3 = "^2.2.1"
yarl = "^1.9.4"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
faker = "^24.2.0"
codecov = "^2.1.13"
coverage = "^7.4.4"
flake8 = "^7.0.0"
black = "^24.3.0"
unify = "^0.5"
autoflake = "^2.3.1"
pytest-cov = "^4.1.0"
flake8-print = "^5.0.0"
tqdm = "^4.66.2"
isort = "^5.13.2"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"